+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/de.json
@@ -1079,7 +1079,7 @@
   "scenes.reminder.back": "Zurückgehen",
   "scenes.reminder.continue": "Weiter",
   "scenes.reset2fa.firststep.reset": "2FA rücksetzen",
-  "scenes.reset2fa.firststep.step1": "Schritt 1 von 3",
+  "scenes.reset2fa.firststep.step1": "Schritt 1 von 2",
   "scenes.reset2fa.firststep.explain": "Füllen Sie das nachfolgende Formular aus, um erneut Zugang zu Ihrer Wallet durch den Reset Ihrer 2-Faktor-Authentifzierung, eingeschränken IP-Adresse und verifizierten E-Mail-Adresse zu erlangen.",
   "scenes.reset2fa.firststep.explain2": "Hinweis: Ihre IP-Adresse und Browser-Informationen werden bei Einreichung aufgezeichnet.",
   "scenes.reset2fa.firststep.firststepform.guid": "Wallet-ID",
@@ -1099,7 +1099,7 @@
   "scenes.reset2fa.secondstep.back": "Zurückgehen",
   "scenes.reset2fa.secondstep.continue": "Weiter",
   "scenes.reset2fa.thirdstep.reset": "Zurücksetzen",
-  "scenes.reset2fa.thirdstep.step3": "Schritt 3 von 3",
+  "scenes.reset2fa.thirdstep.step3": "Schritt 2 von 2",
   "scenes.reset2fa.thirdstep.captcha": "CAPTCHA",
   "scenes.reset2fa.thirdstep.back": "Zurückgehen",
   "scenes.reset2fatoken.error": "<span>Oh-oh. Etwas hat nicht geklappt. Fehler: </span> {error}",
