+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/el.json
@@ -1079,7 +1079,7 @@
   "scenes.reminder.back": "Μετάβαση Πίσω",
   "scenes.reminder.continue": "Συνέχεια",
   "scenes.reset2fa.firststep.reset": "Επαναφορά πιστοποίησης ταυτότητας δύο παραγόντων",
-  "scenes.reset2fa.firststep.step1": "Βήμα 1 από 3",
+  "scenes.reset2fa.firststep.step1": "Βήμα 1 από 2",
   "scenes.reset2fa.firststep.explain": "Συμπληρώστε την πιο κάτω φόρμα για να αποκτήσετε ξανά πρόσβαση στο πορτοφόλι σας με επαναφορά πιστοποίησης ταυτότητας δύο παραγόντων, με περιορισμό IP και επαληθευμένο email.",
   "scenes.reset2fa.firststep.explain2": "Σημείωση: Η διεύθυνση IP και οι πληροφορίες του προγράμματος περιήγησης θα καταγραφούν με την υποβολή.",
   "scenes.reset2fa.firststep.firststepform.guid": "Αναγνωριστικό Πορτοφολιού",
@@ -1099,7 +1099,7 @@
   "scenes.reset2fa.secondstep.back": "Μετάβαση Πίσω",
   "scenes.reset2fa.secondstep.continue": "Συνέχεια",
   "scenes.reset2fa.thirdstep.reset": "Επαναφορά",
-  "scenes.reset2fa.thirdstep.step3": "Βήμα 3 από 3",
+  "scenes.reset2fa.thirdstep.step3": "Βήμα 2 από 2",
   "scenes.reset2fa.thirdstep.captcha": "Κωδικός Captcha",
   "scenes.reset2fa.thirdstep.back": "Μετάβαση Πίσω",
   "scenes.reset2fatoken.error": "<span>Ωχ. Κάτι πήγε στραβά. Σφάλμα: </span> {error}",
