+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/en.json
@@ -1206,7 +1206,7 @@
   "scenes.reminder.login": "Continue to Login",
   "scenes.reminder.remind": "Remind Me",
   "scenes.reset2fa.firststep.reset": "Reset 2FA",
-  "scenes.reset2fa.firststep.step1": "Step 1 of 3",
+  "scenes.reset2fa.firststep.step1": "Step 1 of 2",
   "scenes.reset2fa.firststep.explain": "Fill out the form below to regain access to your wallet by resetting your 2FA, restricted IP, and verified email.",
   "scenes.reset2fa.firststep.explain2": "Note: Your IP address and browser information will be recorded upon submission.",
   "scenes.reset2fa.firststep.firststepform.guid": "Wallet Identifier",
@@ -1226,7 +1226,7 @@
   "scenes.reset2fa.secondstep.back": "Go Back",
   "scenes.reset2fa.secondstep.continue": "Continue",
   "scenes.reset2fa.thirdstep.reset": "Reset",
-  "scenes.reset2fa.thirdstep.step3": "Step 3 of 3",
+  "scenes.reset2fa.thirdstep.step3": "Step 2 of 2",
   "scenes.reset2fa.thirdstep.captcha": "Captcha",
   "scenes.reset2fa.thirdstep.back": "Go Back",
   "scenes.reset2fatoken.error": "<span>Uh Oh. Something went wrong. Error: </span> {error}",
