+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/es.json
@@ -1079,7 +1079,7 @@
   "scenes.reminder.back": "Volver",
   "scenes.reminder.continue": "Continuar",
   "scenes.reset2fa.firststep.reset": "Restablecer autenticación 2FA",
-  "scenes.reset2fa.firststep.step1": "Paso 1 de 3",
+  "scenes.reset2fa.firststep.step1": "Paso 1 de 2",
   "scenes.reset2fa.firststep.explain": "Rellene el siguiente formulario para volver a tener acceso a su billetera restableciendo la autenticación 2FA, IP restringida o por correo electrónico verificado.",
   "scenes.reset2fa.firststep.explain2": "Nota: Su dirección IP y la información de su navegador quedarán registrados al enviarlo.",
   "scenes.reset2fa.firststep.firststepform.guid": "Identificador de billetera",
@@ -1099,7 +1099,7 @@
   "scenes.reset2fa.secondstep.back": "Volver",
   "scenes.reset2fa.secondstep.continue": "Continuar",
   "scenes.reset2fa.thirdstep.reset": "Restablecer",
-  "scenes.reset2fa.thirdstep.step3": "Paso 3 de 3",
+  "scenes.reset2fa.thirdstep.step3": "Paso 2 de 2",
   "scenes.reset2fa.thirdstep.captcha": "Captcha",
   "scenes.reset2fa.thirdstep.back": "Volver",
   "scenes.reset2fatoken.error": "<span>¡Vaya! Algo ha salido mal. Error: </span> {error}",
