+++ packages/blockchain-wallet-v4-frontend/src/assets/locales/ru.json
@@ -1079,7 +1079,7 @@
   "scenes.reminder.back": "Вернуться",
   "scenes.reminder.continue": "Продолжить",
   "scenes.reset2fa.firststep.reset": "Восстановить 2ФА",
-  "scenes.reset2fa.firststep.step1": "Шаг 1 из 3",
+  "scenes.reset2fa.firststep.step1": "Шаг 1 из 2",
   "scenes.reset2fa.firststep.explain": "Заполните форму ниже, чтобы восстановить доступ к вашему кошельку, сбросив 2ФА, ограниченный IP-адрес и верифицированный адрес эл. почты.",
   "scenes.reset2fa.firststep.explain2": "Примечание: при отправке ваш IP-адрес и информация о браузере будут записаны.",
   "scenes.reset2fa.firststep.firststepform.guid": "Идентификатор кошелька",
@@ -1099,7 +1099,7 @@
   "scenes.reset2fa.secondstep.back": "Вернуться",
   "scenes.reset2fa.secondstep.continue": "Продолжить",
   "scenes.reset2fa.thirdstep.reset": "Сброс",
-  "scenes.reset2fa.thirdstep.step3": "Шаг 3 из 3",
+  "scenes.reset2fa.thirdstep.step3": "Шаг 2 из 2",
   "scenes.reset2fa.thirdstep.captcha": "Капча",
   "scenes.reset2fa.thirdstep.back": "Вернуться",
   "scenes.reset2fatoken.error": "<span>Хм. Что-то пошло не так. Ошибка: </span> {error}",
