+++ packages/blockchain-wallet-v4-frontend/src/scenes/Reset2FA/SecondStep/index.js
@@ -1,5 +1,10 @@
 import React from 'react'
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
+import { formValueSelector } from 'redux-form'
+
 import SecondStep from './template'
+import { actions, selectors } from 'data'
 
 class SecondStepContainer extends React.PureComponent {
   constructor (props) {
@@ -8,12 +13,34 @@ class SecondStepContainer extends React.PureComponent {
   }
 
   onSubmit () {
-    this.props.nextStep()
+    const { guid, email, newEmail, secretPhrase, message, code, captcha } = this.props
+    const { sessionToken } = captcha.getOrElse({})
+
+    this.props.authActions.reset2fa(guid, email, newEmail, secretPhrase, message, code, sessionToken)
   }
 
   render () {
-    return <SecondStep {...this.props} onSubmit={this.onSubmit} />
+    const { data } = this.props
+    let busy = data.cata({ Success: () => { this.props.nextStep(); return false }, Failure: () => false, Loading: () => true, NotAsked: () => false })
+
+    return <SecondStep {...this.props} fetchNewCaptcha={this.fetchNewCaptcha} onSubmit={this.onSubmit} busy={busy} />
   }
 }
 
-export default SecondStepContainer
+const mapStateToProps = (state) => ({
+  guid: formValueSelector('reset2FA')(state, 'guid'),
+  email: formValueSelector('reset2FA')(state, 'email'),
+  newEmail: formValueSelector('reset2FA')(state, 'newEmail'),
+  secretPhrase: formValueSelector('reset2FA')(state, 'secretPhrase'),
+  message: formValueSelector('reset2FA')(state, 'message'),
+  code: formValueSelector('reset2FA')(state, 'code'),
+  captcha: selectors.core.data.misc.getCaptcha(state),
+  data: selectors.auth.getReset2fa(state)
+})
+
+const mapDispatchToProps = (dispatch) => ({
+  authActions: bindActionCreators(actions.auth, dispatch),
+  alertActions: bindActionCreators(actions.alerts, dispatch)
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(SecondStepContainer)
