+++ packages/blockchain-wallet-v4-frontend/src/scenes/Reset2FA/index.js
@@ -4,7 +4,6 @@ import wizardProvider from 'providers/WizardProvider'
 import FirstStep from './FirstStep'
 import SecondStep from './SecondStep'
 import ThirdStep from './ThirdStep'
-import FourthStep from './FourthStep'
 
 class Reset2FAContainer extends React.PureComponent {
   componentWillMount () {
@@ -18,10 +17,9 @@ class Reset2FAContainer extends React.PureComponent {
       case 1: return <FirstStep {...rest} />
       case 2: return <SecondStep {...rest} />
       case 3: return <ThirdStep {...rest} />
-      case 4: return <FourthStep {...rest} />
       default: return <FirstStep {...rest} />
     }
   }
 }
 
-export default wizardProvider('reset2FA', 4)(Reset2FAContainer)
+export default wizardProvider('reset2FA', 3)(Reset2FAContainer)
